Renamed module `readers` to `io`.
